requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sqlite3-python==1.0.0
schedule==1.2.0
smtplib==0.0.1
email-validator==2.1.0
pytz==2023.3
playwright==1.40.0
aiohttp==3.9.1
cryptography==41.0.7
pydantic==2.5.0
python-dateutil==2.8.2
